Sync API simplified
