+++ src/main/kotlin/com/cognifide/gradle/aem/deploy/DeployTask.kt
@@ -16,11 +16,7 @@ open class DeployTask : SyncTask() {
 
     @TaskAction
     fun deploy() {
-        synchronizeInstances({ sync ->
-            val packagePath = uploadPackage(determineLocalPackage(), sync).path
-
-            installPackage(packagePath, sync)
-        })
+        synchronizeInstances({ it.deployPackage() })
     }
 
 }
