+++ src/main/kotlin/com/cognifide/gradle/aem/deploy/InstallTask.kt
@@ -16,9 +16,7 @@ open class InstallTask : SyncTask() {
 
     @TaskAction
     fun install() {
-        synchronizeInstances({ sync ->
-            installPackage(determineRemotePackagePath(sync), sync)
-        })
+        synchronizeInstances({ it.installPackage() })
     }
 
 }
